migrate queries to use phlexi form
